+++ app/api/routes/auth.py
@@ -1,16 +1,12 @@
 from fastapi import APIRouter
 
+from app.api.deps import auth_backend, fastapi_users
 from app.schemas import UserCreate, UserRead
-from app.users import auth_backend, fastapi_users
 
 router = APIRouter()
 
 # Authentication routes
-router.include_router(
-    fastapi_users.get_auth_router(auth_backend),
-    prefix="/jwt",
-    tags=["auth"]
-)
+router.include_router(fastapi_users.get_auth_router(auth_backend), prefix="/jwt", tags=["auth"])
 
 router.include_router(
     fastapi_users.get_register_router(UserRead, UserCreate),
