--- main.py
@@ -1,4 +1,0 @@
-import uvicorn
-
-if __name__ == "__main__":
-    uvicorn.run("app.app:app", host="0.0.0.0", log_level="info")
